--- a/internal/parallels/parallels.go
+++ b/internal/parallels/parallels.go
@@ -1,19 +1,32 @@
 package parallels
 
-import "sync"
+import (
+	"runtime"
+	"sync"
+)
 
-func Parallel(from, to, n int, f func(from, to int)) {
+// Parallel runs f in parallel for i in [from, to).
+func Parallel(from, to int, f func(i int)) {
+	parallel(from, to, runtime.NumCPU(), f)
+}
+
+func parallel(from, to, n int, f func(i int)) {
 	if n <= 1 {
-		f(from, to)
+		for i := from; i < to; i++ {
+			f(i)
+		}
 		return
 	}
+
 	var wg sync.WaitGroup
-	wg.Add(n)
-	step := (to - from + n - 1) / n
+	step := max((to-from+n-1)/n, 1)
 	for i := from; i < to; i += step {
+		wg.Add(1)
 		go func(from, to int) {
 			defer wg.Done()
-			f(from, to)
+			for i := from; i < to; i++ {
+				f(i)
+			}
 		}(i, min(i+step, to))
 	}
 	wg.Wait()