--- a/internal/parallels/parallels_test.go
+++ b/internal/parallels/parallels_test.go
@@ -7,8 +7,27 @@ import (
 	"testing"
 )
 
+// Experimental function for comparing performance.
+func parallel1(from, to, n int, f func(from, to int)) {
+	if n <= 1 {
+		f(from, to)
+		return
+	}
+	var wg sync.WaitGroup
+	step := max((to-from+n-1)/n, 1)
+	for i := from; i < to; i += step {
+		wg.Add(1)
+		go func(from, to int) {
+			defer wg.Done()
+			f(from, to)
+		}(i, min(i+step, to))
+	}
+	wg.Wait()
+}
+
+// Experimental function for comparing performance.
+// base on https://github.com/disintegration/imaging/blob/d40f48ce0f098c53ab1fcd6e0e402da682262da5/utils.go#L20-L50
 func parallel2(start, stop, n int, fn func(<-chan int)) {
-	// https://github.com/disintegration/imaging/blob/d40f48ce0f098c53ab1fcd6e0e402da682262da5/utils.go#L20-L50
 	count := stop - start
 	c := make(chan int, count)
 	for i := start; i < stop; i++ {
@@ -27,11 +46,8 @@ func parallel2(start, stop, n int, fn func(<-chan int)) {
 	wg.Wait()
 }
 
-// RunWorkers executes the specified worker function using n goroutines, passing
-// each a workerNum from 0-n and a workerCount of n. This function returns after
-// all workers have run to completion.
-//
-// https://github.com/mandykoh/go-parallel/blob/467982c50985631149f6cb6d7d5a34541703b2f4/parallel.go#L7-L22
+// Experimental function for comparing performance.
+// based on https://github.com/mandykoh/go-parallel/blob/467982c50985631149f6cb6d7d5a34541703b2f4/parallel.go#L7-L22
 func RunWorkers(n int, worker func(workerNum, workerCount int)) {
 	allDone := sync.WaitGroup{}
 	allDone.Add(n)
@@ -46,6 +62,7 @@ func RunWorkers(n int, worker func(workerNum, workerCount int)) {
 	allDone.Wait()
 }
 
+// Experimental function for comparing performance.
 func parallel4(from, to, n int, f func(i int)) {
 	if n <= 1 {
 		for i := from; i < to; i++ {
@@ -53,10 +70,11 @@ func parallel4(from, to, n int, f func(i int)) {
 		}
 		return
 	}
+
 	var wg sync.WaitGroup
-	wg.Add(n)
-	step := (to - from + n - 1) / n
+	step := max((to-from+n-1)/n, 1)
 	for i := from; i < to; i += step {
+		wg.Add(1)
 		go func(from, to int) {
 			defer wg.Done()
 			for i := from; i < to; i++ {
@@ -75,7 +93,25 @@ func BenchmarkParallel(b *testing.B) {
 		n := n
 		b.Run(fmt.Sprintf("%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				Parallel(dst.Bounds().Min.Y, dst.Bounds().Max.Y, n, func(from, to int) {
+				parallel(dst.Bounds().Min.Y, dst.Bounds().Max.Y, n, func(y int) {
+					for x := 0; x < 512; x++ {
+						dst.SetNRGBA64(x, y, src.NRGBA64At(x, y))
+					}
+				})
+			}
+		})
+	}
+}
+
+func BenchmarkParallel1(b *testing.B) {
+	src := image.NewNRGBA64(image.Rect(0, 0, 512, 512))
+	dst := image.NewNRGBA64(image.Rect(0, 0, 512, 512))
+
+	for n := 1; n <= 16; n++ {
+		n := n
+		b.Run(fmt.Sprintf("%d", n), func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				parallel1(dst.Bounds().Min.Y, dst.Bounds().Max.Y, n, func(from, to int) {
 					for y := from; y < to; y++ {
 						for x := 0; x < 512; x++ {
 							dst.SetNRGBA64(x, y, src.NRGBA64At(x, y))