--- a/srgb/srgb.go
+++ b/srgb/srgb.go
@@ -8,13 +8,15 @@ package srgb
 import (
 	"image"
 	"image/color"
+
+	"github.com/shogo82148/go-imaging/internal/parallels"
 )
 
 // Decode decodes an sRGB color encoded image to a linear color image.
 func Decode(img image.Image) *image.NRGBA64 {
 	bounds := img.Bounds()
 	ret := image.NewNRGBA64(bounds)
-	for y := bounds.Min.Y; y < bounds.Max.Y; y++ {
+	parallels.Parallel(bounds.Min.Y, bounds.Max.Y, func(y int) {
 		for x := bounds.Min.X; x < bounds.Max.X; x++ {
 			c := img.At(x, y)
 			c = color.NRGBA64Model.Convert(c)
@@ -26,7 +28,7 @@ func Decode(img image.Image) *image.NRGBA64 {
 				A: rgba.A,
 			})
 		}
-	}
+	})
 
 	return ret
 }
@@ -35,7 +37,7 @@ func Decode(img image.Image) *image.NRGBA64 {
 func Encode(img *image.NRGBA64) *image.NRGBA64 {
 	bounds := img.Bounds()
 	ret := image.NewNRGBA64(bounds)
-	for y := bounds.Min.Y; y < bounds.Max.Y; y++ {
+	parallels.Parallel(bounds.Min.Y, bounds.Max.Y, func(y int) {
 		for x := bounds.Min.X; x < bounds.Max.X; x++ {
 			rgba := img.NRGBA64At(x, y)
 			ret.SetRGBA64(x, y, color.RGBA64{
@@ -45,7 +47,7 @@ func Encode(img *image.NRGBA64) *image.NRGBA64 {
 				A: rgba.A,
 			})
 		}
-	}
+	})
 
 	return ret
 }