--- a/srgb/srgb_test.go
+++ b/srgb/srgb_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestDecode(t *testing.T) {
-	input := image.NewRGBA64(image.Rect(0, 0, 10, 1))
+	input := image.NewNRGBA64(image.Rect(0, 0, 10, 1))
 	input.SetRGBA64(0, 0, color.RGBA64{0x0000, 0x0000, 0x000, 0x0000})
 	input.SetRGBA64(1, 0, color.RGBA64{0xffff, 0x0000, 0x000, 0xffff})
 	input.SetRGBA64(2, 0, color.RGBA64{0x8000, 0x0000, 0x000, 0xffff})
@@ -21,6 +21,13 @@ func TestDecode(t *testing.T) {
 	}
 	if got, want := output.NRGBA64At(2, 0), (color.NRGBA64{0x36cc, 0x0000, 0x000, 0xffff}); got != want {
 		t.Errorf("unexpected color: got %v, want %v", got, want)
+	}
+}
+
+func BenchmarkDecode(b *testing.B) {
+	input := image.NewNRGBA64(image.Rect(0, 0, 512, 512))
+	for i := 0; i < b.N; i++ {
+		Decode(input)
 	}
 }
 
@@ -39,5 +46,12 @@ func TestEncode(t *testing.T) {
 	}
 	if got, want := output.NRGBA64At(2, 0), (color.NRGBA64{0x8000, 0x0000, 0x000, 0xffff}); got != want {
 		t.Errorf("unexpected color: got %v, want %v", got, want)
+	}
+}
+
+func BenchmarkEncode(b *testing.B) {
+	input := image.NewNRGBA64(image.Rect(0, 0, 512, 512))
+	for i := 0; i < b.N; i++ {
+		Encode(input)
 	}
 }