Parallelize image processing