@shogo82148 2023-08-23
Parallel4 looks good choice.

```
shogo@ichinoseshougonoMacBook-Pro parallels % benchstat -col .name .new.txt
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/go-imaging/internal/parallels
        │  Parallel   │              Parallel2              │             Parallel3              │             Parallel4              │
        │   sec/op    │   sec/op     vs base                │   sec/op     vs base               │   sec/op     vs base               │
*/1-10    1.203m ± 1%   1.208m ± 1%        ~ (p=0.393 n=10)   1.264m ± 1%  +5.04% (p=0.000 n=10)   1.212m ± 2%       ~ (p=0.393 n=10)
*/2-10    651.1µ ± 0%   609.2µ ± 0%   -6.44% (p=0.000 n=10)   658.9µ ± 1%  +1.20% (p=0.019 n=10)   633.5µ ± 1%  -2.71% (p=0.000 n=10)
*/3-10    471.1µ ± 1%   437.5µ ± 1%   -7.13% (p=0.000 n=10)   468.8µ ± 4%       ~ (p=0.739 n=10)   450.7µ ± 0%  -4.32% (p=0.000 n=10)
*/4-10    374.6µ ± 1%   331.8µ ± 0%  -11.43% (p=0.000 n=10)   369.2µ ± 1%  -1.46% (p=0.000 n=10)   360.9µ ± 0%  -3.67% (p=0.000 n=10)
*/5-10    324.7µ ± 4%   358.5µ ± 0%  +10.43% (p=0.000 n=10)   311.1µ ± 0%  -4.17% (p=0.000 n=10)   305.0µ ± 0%  -6.07% (p=0.000 n=10)
*/6-10    285.9µ ± 1%   354.4µ ± 0%  +23.97% (p=0.000 n=10)   285.2µ ± 0%       ~ (p=0.063 n=10)   281.7µ ± 0%  -1.45% (p=0.000 n=10)
*/7-10    288.6µ ± 2%   352.4µ ± 0%  +22.13% (p=0.000 n=10)   286.7µ ± 0%  -0.64% (p=0.000 n=10)   283.7µ ± 0%  -1.67% (p=0.000 n=10)
*/8-10    282.4µ ± 0%   346.3µ ± 0%  +22.62% (p=0.000 n=10)   285.5µ ± 0%  +1.09% (p=0.000 n=10)   278.3µ ± 0%  -1.44% (p=0.000 n=10)
*/9-10    282.9µ ± 0%   358.2µ ± 0%  +26.60% (p=0.000 n=10)   278.1µ ± 0%  -1.71% (p=0.000 n=10)   275.7µ ± 0%  -2.54% (p=0.000 n=10)
*/10-10   279.6µ ± 0%   364.4µ ± 0%  +30.31% (p=0.000 n=10)   271.4µ ± 0%  -2.94% (p=0.000 n=10)   270.6µ ± 0%  -3.23% (p=0.000 n=10)
*/11-10   286.2µ ± 0%   364.9µ ± 0%  +27.51% (p=0.000 n=10)   283.2µ ± 0%  -1.03% (p=0.000 n=10)   282.0µ ± 0%  -1.47% (p=0.000 n=10)
*/12-10   279.0µ ± 0%   365.4µ ± 0%  +30.94% (p=0.000 n=10)   279.7µ ± 1%       ~ (p=0.184 n=10)   273.9µ ± 0%  -1.83% (p=0.000 n=10)
*/13-10   274.0µ ± 2%   364.7µ ± 1%  +33.07% (p=0.000 n=10)   275.4µ ± 0%       ~ (p=0.118 n=10)   270.2µ ± 0%  -1.42% (p=0.001 n=10)
*/14-10   269.0µ ± 0%   366.0µ ± 0%  +36.07% (p=0.000 n=10)   271.5µ ± 0%  +0.92% (p=0.000 n=10)   265.2µ ± 0%  -1.42% (p=0.000 n=10)
*/15-10   264.6µ ± 0%   366.3µ ± 0%  +38.45% (p=0.000 n=10)   266.2µ ± 1%  +0.60% (p=0.023 n=10)   261.7µ ± 0%  -1.10% (p=0.000 n=10)
*/16-10   262.5µ ± 0%   366.7µ ± 0%  +39.71% (p=0.000 n=10)   263.6µ ± 1%  +0.42% (p=0.001 n=10)   259.1µ ± 0%  -1.28% (p=0.001 n=10)
geomean   341.2µ        404.7µ       +18.58%                  340.6µ       -0.18%                  333.8µ       -2.19%
```